Move more types
